+++ app/Http/Controllers/Admin/UrbanGoodz/UrbanGoodzBusinessTypeController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers\Admin\UrbanGoodz;
 
+use App\CentralLogics\Helpers;
 use App\Http\Controllers\Controller;
 use App\Models\UrbanGoodzBusinessType;
 use App\Models\UrbanGoodzCapability;
+use Brian2694\Toastr\Facades\Toastr;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 
@@ -33,8 +35,12 @@ class UrbanGoodzBusinessTypeController extends Controller
 
     public function store(Request $request)
     {
+        if (!Helpers::module_permission_check('urban_goodz_business_type_manage')) {
+            Toastr::error(translate('messages.access_denied'));
+            return back();
+        }
+
         $data = $request->validate([
-            'slug' => ['required', 'string', 'max:100', 'alpha_dash', Rule::unique('urban_goodz_business_types', 'slug')],
             'name' => ['required', 'string', 'max:255'],
             'description' => ['nullable', 'string'],
             'icon' => ['nullable', 'string', 'max:100'],
@@ -58,6 +64,11 @@ class UrbanGoodzBusinessTypeController extends Controller
 
     public function update(Request $request, $id)
     {
+        if (!Helpers::module_permission_check('urban_goodz_business_type_manage')) {
+            Toastr::error(translate('messages.access_denied'));
+            return back();
+        }
+
         $type = UrbanGoodzBusinessType::findOrFail($id);
 
         $data = $request->validate([
@@ -79,6 +90,11 @@ class UrbanGoodzBusinessTypeController extends Controller
 
     public function destroy($id)
     {
+        if (!Helpers::module_permission_check('urban_goodz_business_type_manage')) {
+            Toastr::error(translate('messages.access_denied'));
+            return back();
+        }
+
         $type = UrbanGoodzBusinessType::findOrFail($id);
         $type->capabilities()->detach();
         $type->delete();
